introduce `ChangeRate`
