+++ src/modals/ManagePool/Forms/ManageCommission/index.tsx
@@ -10,7 +10,6 @@ import {
   ModalWarnings,
 } from '@polkadot-cloud/react';
 import BigNumber from 'bignumber.js';
-import { intervalToDuration } from 'date-fns';
 import { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useApi } from 'contexts/Api';
@@ -18,7 +17,6 @@ import { useHelp } from 'contexts/Help';
 import { useActivePools } from 'contexts/Pools/ActivePools';
 import { useBondedPools } from 'contexts/Pools/BondedPools';
 import { usePoolsConfig } from 'contexts/Pools/PoolsConfig';
-import { MinDelayInput } from 'library/Form/MinDelayInput';
 import { Warning } from 'library/Form/Warning';
 import { useBatchCall } from 'library/Hooks/useBatchCall';
 import { useSignerWarnings } from 'library/Hooks/useSignerWarnings';
@@ -27,12 +25,10 @@ import { SubmitTx } from 'library/SubmitTx';
 import 'rc-slider/assets/index.css';
 import { useOverlay } from '@polkadot-cloud/react/hooks';
 import { useActiveAccounts } from 'contexts/ActiveAccounts';
-import { StyledSlider } from 'library/StyledSlider';
-import { SliderWrapper } from '../../Wrappers';
-import type { ChangeRateInput } from '../types';
 import { usePoolCommission } from './provider';
 import { CommissionCurrent } from './CommissionCurrent';
 import { MaxCommission } from './MaxCommission';
+import { ChangeRate } from './ChangeRate';
 
 export const ManageCommission = ({
   setSection,
@@ -40,7 +36,7 @@ export const ManageCommission = ({
 }: any) => {
   const { t } = useTranslation('modals');
   const { openHelp } = useHelp();
-  const { api, consts } = useApi();
+  const { api } = useApi();
   const { stats } = usePoolsConfig();
   const { newBatchCall } = useBatchCall();
   const { activeAccount } = useActiveAccounts();
@@ -49,7 +45,6 @@ export const ManageCommission = ({
   const { isOwner, selectedActivePool } = useActivePools();
   const { getBondedPool, updateBondedPools } = useBondedPools();
   const {
-    setChangeRate,
     getInitial,
     getCurrent,
     getEnabled,
@@ -64,71 +59,12 @@ export const ManageCommission = ({
   const maxCommission = getCurrent('max_commission');
   const changeRate = getCurrent('change_rate');
 
-  const { expectedBlockTime } = consts;
   const { globalMaxCommission } = stats;
   const poolId = selectedActivePool?.id || 0;
   const bondedPool = getBondedPool(poolId);
 
-  // Convert a block number into an estimated change rate duration.
-  const minDelayToInput = (delay: number) => {
-    const milliseconds = expectedBlockTime.multipliedBy(delay);
-    const end = milliseconds.isZero()
-      ? 0
-      : milliseconds.integerValue().toNumber();
-
-    const { years, months, days, hours, minutes } = intervalToDuration({
-      start: 0,
-      end,
-    });
-
-    return {
-      years: years || 0,
-      months: months || 0,
-      days: days || 0,
-      hours: hours || 0,
-      minutes: minutes || 0,
-    };
-  };
-
-  const inputToMinDelay = (input: ChangeRateInput) => {
-    const { years, months, days, hours, minutes } = input;
-
-    // calculate number of seconds from changeRateInput
-    const yearsSeconds = new BigNumber(years).multipliedBy(31536000);
-    const monthsSeconds = new BigNumber(months).multipliedBy(2628288);
-    const daysSeconds = new BigNumber(days).multipliedBy(86400);
-    const hoursSeconds = new BigNumber(hours).multipliedBy(3600);
-    const minutesSeconds = new BigNumber(minutes).multipliedBy(60);
-
-    return yearsSeconds
-      .plus(monthsSeconds)
-      .plus(daysSeconds)
-      .plus(hoursSeconds)
-      .plus(minutesSeconds)
-      .dividedBy(expectedBlockTime.dividedBy(1000))
-      .integerValue()
-      .toNumber();
-  };
-
-  // Store the change rate value in input format.
-  const [changeRateInput, setChangeRateInput] = useState<ChangeRateInput>(
-    minDelayToInput(changeRate.minDelay)
-  );
-
   // Valid to submit transaction
   const [valid, setValid] = useState<boolean>(false);
-
-  const handleChangeRateInput = (field: string, value: number) => {
-    const newChangeRateInput = {
-      ...changeRateInput,
-      [field]: value,
-    };
-    setChangeRateInput(newChangeRateInput);
-    setChangeRate({
-      ...changeRate,
-      minDelay: inputToMinDelay(newChangeRateInput),
-    });
-  };
 
   const hasCurrentCommission = payee && commission !== 0;
   const commissionCurrent = () => {
@@ -136,7 +72,6 @@ export const ManageCommission = ({
   };
 
   // Monitor when input items change.
-
   const commissionUpdated = commission !== getInitial('commission');
 
   const changeRateUpdated =
@@ -147,11 +82,6 @@ export const ManageCommission = ({
       JSON.stringify(changeRate) !==
         JSON.stringify(getInitial('change_rate'))) ||
     (!hasValue('change_rate') && getEnabled('change_Rate'));
-
-  const maxIncreaseUpdated =
-    changeRate.maxIncrease !== getInitial('change_rate').maxIncrease;
-  const minDelayUpdated =
-    changeRate.minDelay !== getInitial('change_rate').minDelay;
 
   // Global form change.
   const noChange =
@@ -310,38 +240,6 @@ export const ManageCommission = ({
     submitExtrinsic.proxySupported
   );
 
-  const maxIncreaseFeedback = (() => {
-    if (!maxIncreaseUpdated) {
-      return undefined;
-    }
-    if (invalidMaxIncrease) {
-      return {
-        text: t('aboveExisting'),
-        label: 'danger',
-      };
-    }
-    return {
-      text: t('updated'),
-      label: 'neutral',
-    };
-  })();
-
-  const minDelayFeedback = (() => {
-    if (!minDelayUpdated) {
-      return undefined;
-    }
-    if (invalidMinDelay) {
-      return {
-        text: t('belowExisting'),
-        label: 'danger',
-      };
-    }
-    return {
-      text: t('updated'),
-      label: 'neutral',
-    };
-  })();
-
   return (
     <>
       <ModalPadding horizontalOnly>
@@ -403,75 +301,11 @@ export const ManageCommission = ({
           }
         />
 
-        {getEnabled('change_rate') && (
-          <SliderWrapper>
-            <div>
-              <h2>{changeRate.maxIncrease}% </h2>
-              <h5 className={maxIncreaseFeedback?.label || 'neutral'}>
-                {!!maxIncreaseFeedback && maxIncreaseFeedback.text}
-              </h5>
-            </div>
-
-            <StyledSlider
-              value={changeRate.maxIncrease}
-              step={0.1}
-              onChange={(val) => {
-                if (typeof val === 'number') {
-                  setChangeRate({
-                    ...changeRate,
-                    maxIncrease: val,
-                  });
-                }
-              }}
-            />
-
-            <h5 style={{ marginTop: '1rem' }}>
-              {t('minDelayBetweenUpdates')}
-              {minDelayFeedback && (
-                <span className={minDelayFeedback?.label || 'neutral'}>
-                  {minDelayFeedback.text}
-                </span>
-              )}
-            </h5>
-            <div className="changeRate">
-              <MinDelayInput
-                initial={changeRateInput.years}
-                field="years"
-                label={t('years')}
-                handleChange={handleChangeRateInput}
-              />
-              <MinDelayInput
-                initial={changeRateInput.months}
-                field="months"
-                label={t('months')}
-                handleChange={handleChangeRateInput}
-              />
-              <MinDelayInput
-                initial={changeRateInput.days}
-                field="days"
-                label={t('days')}
-                handleChange={handleChangeRateInput}
-              />
-              <MinDelayInput
-                initial={changeRateInput.hours}
-                field="hours"
-                label={t('hours')}
-                handleChange={handleChangeRateInput}
-              />
-              <MinDelayInput
-                initial={changeRateInput.minutes}
-                field="minutes"
-                label={t('minutes')}
-                handleChange={handleChangeRateInput}
-              />
-            </div>
-            <p>
-              {t('thisMinimumDelay', {
-                count: changeRate.minDelay,
-              })}
-            </p>
-          </SliderWrapper>
-        )}
+        {/* TODO: spread these commission meta values */}
+        <ChangeRate
+          invalidMaxIncrease={invalidMaxIncrease}
+          invalidMinDelay={invalidMinDelay}
+        />
       </ModalPadding>
       <SubmitTx
         valid={valid}
